modified card effect function
