+++ client/src/battle/cardEffect/sniper/attack.ts
@@ -1,14 +1,31 @@
 import { CardEffectProps } from '../../../types/battle/cardEffect'
-import { oneAttack, allAttack } from '../../../common/cardEffect'
+import { playerAttack } from '../../../battle/player'
+import { setDamage } from '../../../battle/enemy'
 
 export const eagred = (props: CardEffectProps): void => {
-  oneAttack(props)
+  if (props.type === "oneAttack") {
+    const { player, enemy, card } = props
+    const damage = playerAttack(player, enemy, card)
+    setDamage(enemy, damage)
+  }
 }
 
 export const lightning = (props: CardEffectProps): void => {
-  allAttack(props)
+  if (props.type === "allAttack") {
+    const { player, enemies, card } = props
+    enemies.forEach(enemy => {
+      const damage = playerAttack(player, enemy, card)
+      setDamage(enemy, damage)
+    })
+  }
 }
 
 export const ibis = (props: CardEffectProps): void => {
-  allAttack(props)
+  if (props.type === "allAttack") {
+    const { player, enemies, card } = props
+    enemies.forEach(enemy => {
+      const damage = playerAttack(player, enemy, card)
+      setDamage(enemy, damage)
+    })
+  }
 }
